Fix build.yml for PRs
The PR workflow was using main's version of the spec rather than the
PR's version to test the build. This was happening because we passed the
raw github URL to bikeshed's remote service to get around a limitation
with using multiple .md files for the spec. This simplifies the build
and Makefile by installing bikeshed as part of the workflow and using a
common build target.