Remove scheduler.yield() parameters
By default, scheduler.yield() continuations inherit the priority and
abort signal from the current task, but this behavior can be overridden
by providing a signal or priority option. But it's not clear if this is
necessary, and there are questions around the expectations of subsequent
continuations (see https://github.com/WICG/scheduling-apis/issues/96).

To simplify the API, for now we're removing the yield() parameters,
instead always using the inherited state. We can revisit this as use
cases arise.